essenceAI/requirements.txt
# essenceAI - B2B Market Intelligence Platform
# Dependencies for the hackathon MVP

streamlit>=1.31.0
llama-index>=0.10.0
llama-index-llms-openai>=0.1.0
llama-index-llms-anthropic>=0.1.0
llama-index-embeddings-openai>=0.1.0
pandas>=2.0.0
plotly>=5.18.0
python-dotenv>=1.0.0
openai>=1.12.0
anthropic>=0.18.0
requests>=2.31.0
tavily-python>=0.3.0
ddgs>=1.0.0

# Optional: Weaviate for cloud vector storage (recommended)
weaviate-client>=3.25.0
# Agent System Dependencies
aiohttp>=3.9.0

# Testing
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
